Moving persistence files into sub-dirs (tindyb, sqlalchemy)
