+++ entropylab/pipeline/api/in_process_param_store.py
@@ -13,7 +13,7 @@ from entropylab.pipeline.api.param_store import (
     Param,
     _ns_to_datetime,
 )
-from entropylab.pipeline.params.persistence.tinydb_persistence import (
+from entropylab.pipeline.params.persistence.tinydb.tinydb_persistence import (
     TinyDBPersistence,
     Commit,
     Metadata,
